requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.47.0
twilio==9.2.3
pyngrok==7.2.0
gtts==2.5.1
requests==2.32.3
streamlit==1.39.0
pandas==2.2.2
fpdf2==2.7.9
pydantic==2.9.2